[NFC] Move pass declarations into separate header files
Prepare for migrating SPIRVWriter to the new PassManager by making
sure the new non-legacy pass declarations are visible outside of their
.cpp file.

(cherry picked from commit 337b171131b13487e280752c92508ded4291aff6)
