+++ lib/SPIRV/OCLTypeToSPIRV.cpp
@@ -327,6 +327,8 @@ Type *OCLTypeToSPIRVBase::getAdaptedType(Value *V) {
 
 } // namespace SPIRV
 
+AnalysisKey OCLTypeToSPIRVPass::Key;
+
 INITIALIZE_PASS(OCLTypeToSPIRVLegacy, "cltytospv", "Adapt OCL types for SPIR-V",
                 false, true)
