+++ lib/SPIRV/SPIRVLowerSaddWithOverflow.cpp
@@ -39,6 +39,7 @@
 //===----------------------------------------------------------------------===//
 #define DEBUG_TYPE "spv-lower-llvm_sadd_with_overflow"
 
+#include "SPIRVLowerSaddWithOverflow.h"
 #include "LLVMSaddWithOverflow.h"
 
 #include "LLVMSPIRVLib.h"
@@ -48,10 +49,8 @@
 #include "llvm/IR/InstVisitor.h"
 #include "llvm/IR/IntrinsicInst.h"
 #include "llvm/IR/Module.h"
-#include "llvm/IR/PassManager.h"
 #include "llvm/IRReader/IRReader.h"
 #include "llvm/Linker/Linker.h"
-#include "llvm/Pass.h"
 #include "llvm/Support/SourceMgr.h"
 
 using namespace llvm;
@@ -59,103 +58,86 @@ using namespace SPIRV;
 
 namespace SPIRV {
 
-class SPIRVLowerSaddWithOverflowBase
-    : public InstVisitor<SPIRVLowerSaddWithOverflowBase> {
-public:
-  SPIRVLowerSaddWithOverflowBase() : Context(nullptr) {}
-  virtual ~SPIRVLowerSaddWithOverflowBase() {}
-  virtual void visitIntrinsicInst(CallInst &I) {
-    IntrinsicInst *II = dyn_cast<IntrinsicInst>(&I);
-    if (!II || II->getIntrinsicID() != Intrinsic::sadd_with_overflow)
-      return;
+void SPIRVLowerSaddWithOverflowBase::visitIntrinsicInst(CallInst &I) {
+  IntrinsicInst *II = dyn_cast<IntrinsicInst>(&I);
+  if (!II || II->getIntrinsicID() != Intrinsic::sadd_with_overflow)
+    return;
 
-    Function *IntrinsicFunc = I.getCalledFunction();
-    assert(IntrinsicFunc && "Missing function");
-    StringRef IntrinsicName = IntrinsicFunc->getName();
-    std::string FuncName = "llvm_sadd_with_overflow_i";
-    if (IntrinsicName.endswith(".i16"))
-      FuncName += "16";
-    else if (IntrinsicName.endswith(".i32"))
-      FuncName += "32";
-    else if (IntrinsicName.endswith(".i64"))
-      FuncName += "64";
-    else {
-      assert(false &&
-             "Unsupported overloading of llvm.sadd.with.overflow intrinsic");
-      return;
-    }
-
-    // Redirect @llvm.sadd.with.overflow.* call to the function we have in
-    // the loaded module @llvm_sadd_with_overflow_*
-    Function *F = Mod->getFunction(FuncName);
-    if (F) { // This function is already linked in.
-      I.setCalledFunction(F);
-      return;
-    }
-    FunctionCallee FC = Mod->getOrInsertFunction(FuncName, I.getFunctionType());
-    I.setCalledFunction(FC);
-
-    // Read LLVM IR with the intrinsic's implementation
-    SMDiagnostic Err;
-    auto MB = MemoryBuffer::getMemBuffer(LLVMSaddWithOverflow);
-    auto SaddWithOverflowModule =
-        parseIR(MB->getMemBufferRef(), Err, *Context,
-                [&](StringRef) { return Mod->getDataLayoutStr(); });
-    if (!SaddWithOverflowModule) {
-      std::string ErrMsg;
-      raw_string_ostream ErrStream(ErrMsg);
-      Err.print("", ErrStream);
-      SPIRVErrorLog EL;
-      EL.checkError(false, SPIRVEC_InvalidLlvmModule, ErrMsg);
-      return;
-    }
-
-    // Link in the intrinsic's implementation.
-    if (!Linker::linkModules(*Mod, std::move(SaddWithOverflowModule),
-                             Linker::LinkOnlyNeeded))
-      TheModuleIsModified = true;
+  Function *IntrinsicFunc = I.getCalledFunction();
+  assert(IntrinsicFunc && "Missing function");
+  StringRef IntrinsicName = IntrinsicFunc->getName();
+  std::string FuncName = "llvm_sadd_with_overflow_i";
+  if (IntrinsicName.endswith(".i16"))
+    FuncName += "16";
+  else if (IntrinsicName.endswith(".i32"))
+    FuncName += "32";
+  else if (IntrinsicName.endswith(".i64"))
+    FuncName += "64";
+  else {
+    assert(false &&
+           "Unsupported overloading of llvm.sadd.with.overflow intrinsic");
+    return;
   }
 
-  bool runLowerSaddWithOverflow(Module &M) {
-    Context = &M.getContext();
-    Mod = &M;
-    visit(M);
+  // Redirect @llvm.sadd.with.overflow.* call to the function we have in
+  // the loaded module @llvm_sadd_with_overflow_*
+  Function *F = Mod->getFunction(FuncName);
+  if (F) { // This function is already linked in.
+    I.setCalledFunction(F);
+    return;
+  }
+  FunctionCallee FC = Mod->getOrInsertFunction(FuncName, I.getFunctionType());
+  I.setCalledFunction(FC);
 
-    verifyRegularizationPass(M, "SPIRVLowerSaddWithOverflow");
-    return TheModuleIsModified;
+  // Read LLVM IR with the intrinsic's implementation
+  SMDiagnostic Err;
+  auto MB = MemoryBuffer::getMemBuffer(LLVMSaddWithOverflow);
+  auto SaddWithOverflowModule =
+      parseIR(MB->getMemBufferRef(), Err, *Context,
+              [&](StringRef) { return Mod->getDataLayoutStr(); });
+  if (!SaddWithOverflowModule) {
+    std::string ErrMsg;
+    raw_string_ostream ErrStream(ErrMsg);
+    Err.print("", ErrStream);
+    SPIRVErrorLog EL;
+    EL.checkError(false, SPIRVEC_InvalidLlvmModule, ErrMsg);
+    return;
   }
 
-private:
-  LLVMContext *Context;
-  Module *Mod;
-  bool TheModuleIsModified = false;
-};
+  // Link in the intrinsic's implementation.
+  if (!Linker::linkModules(*Mod, std::move(SaddWithOverflowModule),
+                           Linker::LinkOnlyNeeded))
+    TheModuleIsModified = true;
+}
 
-class SPIRVLowerSaddWithOverflowPass
-    : public llvm::PassInfoMixin<SPIRVLowerSaddWithOverflowPass>,
-      public SPIRVLowerSaddWithOverflowBase {
-public:
-  llvm::PreservedAnalyses run(llvm::Module &M,
-                              llvm::ModuleAnalysisManager &MAM) {
-    return runLowerSaddWithOverflow(M) ? llvm::PreservedAnalyses::none()
-                                       : llvm::PreservedAnalyses::all();
-  }
-};
+bool SPIRVLowerSaddWithOverflowBase::runLowerSaddWithOverflow(Module &M) {
+  Context = &M.getContext();
+  Mod = &M;
+  visit(M);
 
-class SPIRVLowerSaddWithOverflowLegacy : public ModulePass,
-                                         public SPIRVLowerSaddWithOverflowBase {
-public:
-  SPIRVLowerSaddWithOverflowLegacy() : ModulePass(ID) {
-    initializeSPIRVLowerSaddWithOverflowLegacyPass(
-        *PassRegistry::getPassRegistry());
-  }
+  verifyRegularizationPass(M, "SPIRVLowerSaddWithOverflow");
+  return TheModuleIsModified;
+}
 
-  bool runOnModule(Module &M) override { return runLowerSaddWithOverflow(M); }
+llvm::PreservedAnalyses
+SPIRVLowerSaddWithOverflowPass::run(llvm::Module &M,
+                                    llvm::ModuleAnalysisManager &MAM) {
+  return runLowerSaddWithOverflow(M) ? llvm::PreservedAnalyses::none()
+                                     : llvm::PreservedAnalyses::all();
+}
 
-  static char ID;
-};
+SPIRVLowerSaddWithOverflowLegacy::SPIRVLowerSaddWithOverflowLegacy()
+    : ModulePass(ID) {
+  initializeSPIRVLowerSaddWithOverflowLegacyPass(
+      *PassRegistry::getPassRegistry());
+}
+
+bool SPIRVLowerSaddWithOverflowLegacy::runOnModule(Module &M) {
+  return runLowerSaddWithOverflow(M);
+}
 
 char SPIRVLowerSaddWithOverflowLegacy::ID = 0;
+
 } // namespace SPIRV
 
 INITIALIZE_PASS(SPIRVLowerSaddWithOverflowLegacy,
